+++ game_patch/graphics/d3d11/gr_d3d11_solid.h
@@ -32,7 +32,7 @@ namespace df::gr::d3d11
         ~SolidRenderer();
         void render_solid(rf::GSolid* solid, rf::GRoom** rooms, int num_rooms);
         void render_movable_solid(rf::GSolid* solid, const rf::Vector3& pos, const rf::Matrix3& orient);
-        void render_sky_room(rf::GRoom *room);
+        void render_sky_room(rf::GRoom *room, rf::Vector3& out_sky_transform_pos, rf::Matrix3& out_sky_transform_orient);
         void render_alpha_detail(rf::GRoom *room, rf::GSolid *solid);
         void render_room_liquid_surface(rf::GSolid* solid, rf::GRoom* room);
         void clear_cache();